+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fixed bug #60 "Reconstruction fails at random" which occurs when the secret is created from a base64 string
+
 ### Removed
 - Removed .NET Core 2.1 (LTS) support